requirements.txt
keras==2.1.5
tensorflow-gpu==1.4.0
Pillow==7.2.0
pygame==2.0.0
pylsd-nova==1.2.0
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyqtgraph==0.11.0
numpy==1.18.5
h5py==2.10.0
opencv-python==4.4.0.42
